Endrer gjennomførBehandlingsløp-funksjonen slik at det dynamisk kjører gjennom alle steg ut fra hva ferdigstill-steg endepunktet sier at neste steg er
