+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/dagligReise/DagligReiseVedtakController.kt
@@ -2,7 +2,10 @@ package no.nav.tilleggsstonader.sak.vedtak.dagligReise
 
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import no.nav.tilleggsstonader.sak.behandling.BehandlingService
+import no.nav.tilleggsstonader.sak.behandling.domain.Behandling
+import no.nav.tilleggsstonader.sak.behandlingsflyt.StegFerdigstiltResponse
 import no.nav.tilleggsstonader.sak.behandlingsflyt.StegService
+import no.nav.tilleggsstonader.sak.behandlingsflyt.tilStegFerdigstiltResponse
 import no.nav.tilleggsstonader.sak.felles.domain.BehandlingId
 import no.nav.tilleggsstonader.sak.tilgang.AuditLoggerEvent
 import no.nav.tilleggsstonader.sak.tilgang.TilgangService
@@ -49,23 +52,19 @@ class DagligReiseVedtakController(
     fun innvilge(
         @PathVariable behandlingId: BehandlingId,
         @RequestBody vedtak: InnvilgelseDagligReiseTsoRequest,
-    ) {
-        lagreVedtak(behandlingId, vedtak)
-    }
+    ): StegFerdigstiltResponse = lagreVedtak(behandlingId, vedtak).tilStegFerdigstiltResponse()
 
     @PostMapping("{behandlingId}/tsr/innvilgelse")
     fun innvilge(
         @PathVariable behandlingId: BehandlingId,
         @RequestBody vedtak: InnvilgelseDagligReiseTsrRequest,
-    ) {
-        lagreVedtak(behandlingId, vedtak)
-    }
+    ): StegFerdigstiltResponse = lagreVedtak(behandlingId, vedtak).tilStegFerdigstiltResponse()
 
     @PostMapping("{behandlingId}/avslag")
     fun avslå(
         @PathVariable behandlingId: BehandlingId,
         @RequestBody vedtak: AvslagDagligReiseDto,
-    ) {
+    ): StegFerdigstiltResponse {
         val stønadsType = behandlingService.hentSaksbehandling(behandlingId).stønadstype
 
         validerGyldigÅrsakAvslag.validerAvslagErGyldig(
@@ -74,7 +73,7 @@ class DagligReiseVedtakController(
             stønadsType,
         )
 
-        lagreVedtak(behandlingId, vedtak)
+        return lagreVedtak(behandlingId, vedtak).tilStegFerdigstiltResponse()
     }
 
     @GetMapping("{behandlingId}")
@@ -92,9 +91,7 @@ class DagligReiseVedtakController(
     fun opphor(
         @PathVariable behandlingId: BehandlingId,
         @RequestBody vedtak: OpphørDagligReiseRequest,
-    ) {
-        lagreVedtak(behandlingId, vedtak)
-    }
+    ): StegFerdigstiltResponse = lagreVedtak(behandlingId, vedtak).tilStegFerdigstiltResponse()
 
     @PostMapping("{behandlingId}/tso/beregn")
     fun beregn(
@@ -154,10 +151,10 @@ class DagligReiseVedtakController(
     private fun lagreVedtak(
         behandlingId: BehandlingId,
         vedtak: VedtakDagligReiseRequest,
-    ) {
+    ): Behandling {
         tilgangService.settBehandlingsdetaljerForRequest(behandlingId)
         tilgangService.validerSkrivetilgangTilBehandling(behandlingId, AuditLoggerEvent.CREATE)
 
-        stegService.håndterSteg(behandlingId, beregnYtelseSteg, vedtak)
+        return stegService.håndterSteg(behandlingId, beregnYtelseSteg, vedtak)
     }
 }
