+++ src/test/kotlin/no/nav/tilleggsstonader/sak/integrasjonstest/GjennomførBehandling.kt
@@ -3,20 +3,17 @@ package no.nav.tilleggsstonader.sak.integrasjonstest
 import io.mockk.every
 import no.nav.tilleggsstonader.kontrakter.felles.JsonMapperProvider.jsonMapper
 import no.nav.tilleggsstonader.kontrakter.felles.Stønadstype
-import no.nav.tilleggsstonader.kontrakter.felles.gjelderDagligReise
 import no.nav.tilleggsstonader.kontrakter.journalpost.Dokumentvariantformat
 import no.nav.tilleggsstonader.kontrakter.journalpost.Journalpost
 import no.nav.tilleggsstonader.libs.test.fnr.FnrGenerator
 import no.nav.tilleggsstonader.sak.IntegrationTest
 import no.nav.tilleggsstonader.sak.behandling.domain.BehandlingÅrsak
 import no.nav.tilleggsstonader.sak.behandling.domain.HenlagtÅrsak
+import no.nav.tilleggsstonader.sak.behandling.dto.BehandlingDto
 import no.nav.tilleggsstonader.sak.behandling.dto.HenlagtDto
 import no.nav.tilleggsstonader.sak.behandling.dto.OpprettBehandlingDto
 import no.nav.tilleggsstonader.sak.behandling.opprettelse.ForenkletBehandlingstype
-import no.nav.tilleggsstonader.sak.behandlingsflyt.StegController
 import no.nav.tilleggsstonader.sak.behandlingsflyt.StegType
-import no.nav.tilleggsstonader.sak.brev.GenererPdfRequest
-import no.nav.tilleggsstonader.sak.felles.domain.BarnId
 import no.nav.tilleggsstonader.sak.felles.domain.BehandlingId
 import no.nav.tilleggsstonader.sak.felles.domain.FagsakId
 import no.nav.tilleggsstonader.sak.infrastruktur.mocks.PdlClientMockConfig
@@ -26,35 +23,16 @@ import no.nav.tilleggsstonader.sak.integrasjonstest.extensions.tasks.kjørTasksK
 import no.nav.tilleggsstonader.sak.integrasjonstest.extensions.tilordneÅpenBehandlingOppgaveForBehandling
 import no.nav.tilleggsstonader.sak.integrasjonstest.testdata.defaultJournalpost
 import no.nav.tilleggsstonader.sak.integrasjonstest.testdata.journalpostSøknadForStønadstype
-import no.nav.tilleggsstonader.sak.integrasjonstest.testdata.tilVedtaksperiodeDagligReiseDto
 import no.nav.tilleggsstonader.sak.opplysninger.ytelse.YtelsePerioderUtil.ytelsePerioderDtoAAP
 import no.nav.tilleggsstonader.sak.util.SøknadBoutgifterUtil.søknadBoutgifter
 import no.nav.tilleggsstonader.sak.util.SøknadDagligReiseUtil.søknadDagligReise
 import no.nav.tilleggsstonader.sak.util.SøknadUtil.barnMedBarnepass
 import no.nav.tilleggsstonader.sak.util.SøknadUtil.søknadskjemaBarnetilsyn
 import no.nav.tilleggsstonader.sak.util.SøknadUtil.søknadskjemaLæremidler
-import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.dto.InnvilgelseTilsynBarnRequest
-import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.dto.OpphørTilsynBarnRequest
-import no.nav.tilleggsstonader.sak.vedtak.boutgifter.dto.InnvilgelseBoutgifterRequest
-import no.nav.tilleggsstonader.sak.vedtak.boutgifter.dto.OpphørBoutgifterRequest
-import no.nav.tilleggsstonader.sak.vedtak.dagligReise.dto.InnvilgelseDagligReiseTsoRequest
-import no.nav.tilleggsstonader.sak.vedtak.dagligReise.dto.InnvilgelseDagligReiseTsrRequest
-import no.nav.tilleggsstonader.sak.vedtak.dagligReise.dto.OpphørDagligReiseRequest
 import no.nav.tilleggsstonader.sak.vedtak.domain.ÅrsakOpphør
 import no.nav.tilleggsstonader.sak.vedtak.dto.VedtaksperiodeDto
-import no.nav.tilleggsstonader.sak.vedtak.læremidler.dto.InnvilgelseLæremidlerRequest
-import no.nav.tilleggsstonader.sak.vedtak.læremidler.dto.OpphørLæremidlerRequest
-import no.nav.tilleggsstonader.sak.vedtak.totrinnskontroll.dto.BeslutteVedtakDto
 import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dagligReise.dto.FaktaDagligReisePrivatBilDto
-import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dagligReise.dto.LagreVilkårDagligReiseDto
-import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dagligReise.dto.VilkårDagligReiseDto
-import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dto.OpprettVilkårDto
-import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.dto.VilkårsvurderingDto
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.SlettVikårperiode
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.dto.VilkårperioderDto
-import org.springframework.test.web.servlet.client.RestTestClient
 import java.time.LocalDate
-import java.util.UUID
 
 data class BehandlingContext(
     val behandlingId: BehandlingId,
@@ -151,50 +129,9 @@ fun IntegrationTest.gjennomførBehandlingsløp(
 
     val testdata = BehandlingTestdataDsl.build(testdataProvider)
 
-    if (tilSteg == StegType.INNGANGSVILKÅR) {
-        return
-    }
-
-    gjennomførInngangsvilkårSteg(testdata, behandlingId)
-
-    if (tilSteg == StegType.VILKÅR) {
-        return
-    }
-
-    if (behandling.stønadstype != Stønadstype.LÆREMIDLER) {
-        gjennomførVilkårSteg(testdata, behandling.id, behandling.stønadstype)
-    }
-
-    if (tilSteg == StegType.BEREGNE_YTELSE) {
-        return
-    }
-
-    gjennomførBeregningSteg(behandling.id, behandling.stønadstype, testdata.vedtak.vedtak)
-
-    if (erRevurderingDagligReiseMedPrivatBil) {
-        gjennomførEkstraStegVedDagligReisePrivatBilRevurdering(behandling.id, tilSteg)
-    }
-
-    if (tilSteg == StegType.SIMULERING) {
-        return
-    }
-
-    gjennomførSimuleringSteg(behandlingId)
-
-    if (tilSteg == StegType.SEND_TIL_BESLUTTER) {
-        return
-    }
-
-    gjennomførSendTilBeslutterSteg(behandlingId)
-
-    if (tilSteg == StegType.BESLUTTE_VEDTAK) {
-        return
-    }
-
-    gjennomførBeslutteVedtakSteg(behandlingId)
-
-    if (tilSteg in setOf(StegType.JOURNALFØR_OG_DISTRIBUER_VEDTAKSBREV, StegType.FERDIGSTILLE_BEHANDLING)) {
-        return
+    var nåværendeSteg = behandling.steg
+    while (nåværendeSteg != tilSteg) {
+        nåværendeSteg = utførStegOgReturnerNesteSteg(nåværendeSteg, behandling, testdata)
     }
 
     // Ferdigstiller behandling
@@ -209,6 +146,27 @@ fun IntegrationTest.gjennomførBehandlingsløp(
         sendInnKjøreliste(kjøreliste, ident)
     }
 }
+
+fun IntegrationTest.utførStegOgReturnerNesteSteg(
+    steg: StegType,
+    behandling: BehandlingDto,
+    testdata: BehandlingTestdataDsl,
+): StegType =
+    when (steg) {
+        StegType.INNGANGSVILKÅR -> gjennomførInngangsvilkårSteg(testdata, behandling.id)
+        StegType.VILKÅR -> gjennomførVilkårSteg(testdata, behandling.id, behandling.stønadstype)
+        StegType.BEREGNE_YTELSE -> gjennomførBeregningSteg(behandling.id, behandling.stønadstype, testdata.vedtak.vedtak)
+        StegType.KJØRELISTE -> gjennomførKjørelisteSteg(behandling.id)
+        StegType.BEREGNING -> gjennomførBeregningStegDagligReise(behandling.id)
+        StegType.SIMULERING -> gjennomførSimuleringSteg(behandling.id)
+        StegType.SEND_TIL_BESLUTTER -> gjennomførSendTilBeslutterSteg(behandling.id)
+        StegType.BESLUTTE_VEDTAK -> gjennomførBeslutteVedtakSteg(behandling.id)
+        StegType.FULLFØR_KJØRELISTE,
+        StegType.JOURNALFØR_OG_DISTRIBUER_VEDTAKSBREV,
+        StegType.FERDIGSTILLE_BEHANDLING,
+        StegType.BEHANDLING_FERDIGSTILT,
+        -> error("$steg kan ikke gjennomføres gjennom api-kall")
+    }
 
 fun IntegrationTest.opprettRevurdering(opprettBehandlingDto: OpprettBehandlingDto): BehandlingId {
     val behandlingId = kall.behandling.opprettRevurdering(opprettBehandlingDto)
@@ -286,121 +244,6 @@ fun IntegrationTest.gjennomførHenleggelse(fraJournalpost: Journalpost = default
     return behandlingId
 }
 
-fun IntegrationTest.gjennomførBeslutteVedtakSteg(behandlingId: BehandlingId) {
-    medBrukercontext(bruker = "nissemor", roller = listOf(rolleConfig.beslutterRolle)) {
-        tilordneÅpenBehandlingOppgaveForBehandling(behandlingId)
-        kall.totrinnskontroll.beslutteVedtak(behandlingId, BeslutteVedtakDto(godkjent = true))
-    }
-    kjørTasksKlareForProsessering()
-}
-
-fun IntegrationTest.gjennomførSendTilBeslutterSteg(behandlingId: BehandlingId) {
-    kall.brev.genererPdf(behandlingId, GenererPdfRequest(MINIMALT_BREV))
-    kall.brevmottakere.hent(behandlingId)
-    kall.totrinnskontroll.sendTilBeslutter(behandlingId)
-    kjørTasksKlareForProsessering()
-}
-
-fun IntegrationTest.gjennomførSimuleringSteg(behandlingId: BehandlingId) {
-    kall.steg.ferdigstill(
-        behandlingId,
-        StegController.FerdigstillStegRequest(
-            steg = StegType.SIMULERING,
-        ),
-    )
-    kjørTasksKlareForProsessering()
-}
-
-fun IntegrationTest.gjennomførKjørelisteSteg(behandlingId: BehandlingId) {
-    kall.steg.ferdigstill(behandlingId, StegController.FerdigstillStegRequest(StegType.KJØRELISTE))
-}
-
-fun IntegrationTest.gjennomførBeregningStegDagligReise(behandlingId: BehandlingId) {
-    kall.steg.ferdigstill(behandlingId, StegController.FerdigstillStegRequest(StegType.BEREGNING))
-}
-
-fun IntegrationTest.gjennomførBeregningSteg(
-    behandlingId: BehandlingId,
-    stønadstype: Stønadstype,
-    opprettVedtak: OpprettVedtak = OpprettInnvilgelse(),
-) {
-    gjennomførBeregningStegKall(behandlingId, stønadstype, opprettVedtak)
-        .expectStatus()
-        .isOk
-}
-
-fun IntegrationTest.gjennomførBeregningStegKall(
-    behandlingId: BehandlingId,
-    stønadstype: Stønadstype,
-    opprettVedtak: OpprettVedtak = OpprettInnvilgelse(),
-): RestTestClient.ResponseSpec =
-    when (opprettVedtak) {
-        is OpprettInnvilgelse -> {
-            val vedtaksperioder =
-                opprettVedtak.vedtaksperioder ?: kall.vedtak
-                    .foreslåVedtaksperioder(behandlingId)
-                    .map { it.tilVedtaksperiodeDto() }
-            kall.vedtak.apiRespons
-                .lagreInnvilgelse(
-                    stønadstype = stønadstype,
-                    behandlingId = behandlingId,
-                    innvilgelseDto =
-                        when (stønadstype) {
-                            Stønadstype.BARNETILSYN -> InnvilgelseTilsynBarnRequest(vedtaksperioder = vedtaksperioder)
-                            Stønadstype.LÆREMIDLER -> InnvilgelseLæremidlerRequest(vedtaksperioder = vedtaksperioder)
-                            Stønadstype.BOUTGIFTER -> InnvilgelseBoutgifterRequest(vedtaksperioder = vedtaksperioder)
-                            Stønadstype.DAGLIG_REISE_TSO -> InnvilgelseDagligReiseTsoRequest(vedtaksperioder = vedtaksperioder)
-                            Stønadstype.DAGLIG_REISE_TSR ->
-                                InnvilgelseDagligReiseTsrRequest(
-                                    vedtaksperioder = vedtaksperioder.tilVedtaksperiodeDagligReiseDto(),
-                                )
-
-                            Stønadstype.REISE_TIL_SAMLING_TSO -> TODO("InnvilgelseReiseTilSamlingTsoRequest")
-                        },
-                )
-        }
-
-        is OpprettAvslag -> TODO()
-        is OpprettOpphør ->
-            kall.vedtak.apiRespons
-                .lagreOpphør(
-                    stønadstype = stønadstype,
-                    behandlingId = behandlingId,
-                    opphørDto =
-                        when (stønadstype) {
-                            Stønadstype.BARNETILSYN ->
-                                OpphørTilsynBarnRequest(
-                                    årsakerOpphør = opprettVedtak.årsaker,
-                                    begrunnelse = opprettVedtak.begrunnelse,
-                                    opphørsdato = opprettVedtak.opphørsdato,
-                                )
-
-                            Stønadstype.LÆREMIDLER ->
-                                OpphørLæremidlerRequest(
-                                    årsakerOpphør = opprettVedtak.årsaker,
-                                    begrunnelse = opprettVedtak.begrunnelse,
-                                    opphørsdato = opprettVedtak.opphørsdato,
-                                )
-
-                            Stønadstype.BOUTGIFTER ->
-                                OpphørBoutgifterRequest(
-                                    årsakerOpphør = opprettVedtak.årsaker,
-                                    begrunnelse = opprettVedtak.begrunnelse,
-                                    opphørsdato = opprettVedtak.opphørsdato,
-                                )
-
-                            Stønadstype.DAGLIG_REISE_TSO, Stønadstype.DAGLIG_REISE_TSR ->
-                                OpphørDagligReiseRequest(
-                                    årsakerOpphør = opprettVedtak.årsaker,
-                                    begrunnelse = opprettVedtak.begrunnelse,
-                                    opphørsdato = opprettVedtak.opphørsdato,
-                                )
-
-                            Stønadstype.REISE_TIL_SAMLING_TSO -> TODO("OpphørReiseTilSamlingTsoRequest")
-                        },
-                )
-    }
-
 sealed interface OpprettVedtak
 
 data class OpprettInnvilgelse(
@@ -414,146 +257,3 @@ data class OpprettOpphør(
     val begrunnelse: String = "annet",
     val opphørsdato: LocalDate,
 ) : OpprettVedtak
-
-private const val MINIMALT_BREV = """SAKSBEHANDLER_SIGNATUR - BREVDATO_PLACEHOLDER - BESLUTTER_SIGNATUR"""
-
-private fun IntegrationTest.gjennomførInngangsvilkårSteg(
-    testdataDsl: BehandlingTestdataDsl,
-    behandlingId: BehandlingId,
-) {
-    // Hentes ikke ut om man ikke trenger
-    val vilkårperioder: VilkårperioderDto by lazy {
-        kall.vilkårperiode.hentForBehandling(behandlingId).vilkårperioder
-    }
-
-    // Opprett aktiviteter
-    testdataDsl.aktivitet.opprettScope
-        .build(behandlingId)
-        .forEach { lagreVilkårperiode ->
-            kall.vilkårperiode
-                .opprett(lagreVilkårperiode)
-                .periode!!
-                .id
-        }
-
-    // Oppretter målgrupper
-    testdataDsl.målgruppe.opprettScope
-        .build(behandlingId)
-        .forEach { lagreVilkårperiode ->
-            kall.vilkårperiode
-                .opprett(lagreVilkårperiode)
-                .periode!!
-                .id
-        }
-
-    // Oppdater aktiviteter
-    testdataDsl.aktivitet.update.forEach { upd ->
-        val (vilkårperiodeId, lagreVilkårperiode) = upd(vilkårperioder.aktiviteter, behandlingId)
-        kall.vilkårperiode.oppdater(
-            vilkårperiodeId = vilkårperiodeId,
-            lagreVilkårperiode = lagreVilkårperiode,
-        )
-    }
-
-    // Slett aktiviteter
-    testdataDsl.aktivitet.delete.forEach { del ->
-        val idOgRequest: Pair<UUID, SlettVikårperiode> = del(vilkårperioder.aktiviteter)
-        kall.vilkårperiode.apiRespons.slett(
-            vilkårperiodeId = idOgRequest.first,
-            slettVikårperiode = idOgRequest.second,
-        )
-    }
-
-    // Oppdater målgruppeer
-    testdataDsl.målgruppe.update.forEach { upd ->
-        val (vilkårperiodeId, lagreVilkårperiode) = upd(vilkårperioder.målgrupper, behandlingId)
-        kall.vilkårperiode.oppdater(
-            vilkårperiodeId = vilkårperiodeId,
-            lagreVilkårperiode = lagreVilkårperiode,
-        )
-    }
-
-    // Slett målgruppeer
-    testdataDsl.målgruppe.delete.forEach { del ->
-        val idOgRequest: Pair<UUID, SlettVikårperiode> = del(vilkårperioder.målgrupper)
-        kall.vilkårperiode.apiRespons.slett(
-            vilkårperiodeId = idOgRequest.first,
-            slettVikårperiode = idOgRequest.second,
-        )
-    }
-
-    kall.steg.ferdigstill(
-        behandlingId,
-        StegController.FerdigstillStegRequest(
-            steg = StegType.INNGANGSVILKÅR,
-        ),
-    )
-    kjørTasksKlareForProsessering()
-}
-
-private fun IntegrationTest.gjennomførVilkårSteg(
-    testdataProvider: BehandlingTestdataDsl,
-    behandlingId: BehandlingId,
-    stønadstype: Stønadstype,
-) {
-    val vilkårDagligReise: List<VilkårDagligReiseDto> by lazy {
-        kall.vilkårDagligReise.hentVilkår(behandlingId)
-    }
-
-    val vilkår: VilkårsvurderingDto by lazy {
-        kall.vilkår.hentVilkår(behandlingId)
-    }
-
-    // Trenger kun hente ut om tilsyn-barn
-    val barnIder: List<BarnId> =
-        stønadstype
-            .takeIf { it == Stønadstype.BARNETILSYN }
-            ?.let { _ -> barnRepository.findByBehandlingId(behandlingId).map { it.id } }
-            ?: emptyList()
-
-    testdataProvider.vilkår.opprettScope
-        .build(
-            behandlingId,
-            barnIder,
-            aktiviteter =
-                kall.vilkårperiode
-                    .hentForBehandling(behandlingId)
-                    .vilkårperioder.aktiviteter,
-        ).forEach {
-            if (stønadstype.gjelderDagligReise()) {
-                kall.vilkårDagligReise.opprettVilkår(behandlingId, it as LagreVilkårDagligReiseDto)
-            } else {
-                kall.vilkår.opprettVilkår(it as OpprettVilkårDto)
-            }
-        }
-
-    if (stønadstype.gjelderDagligReise()) {
-        val aktiviteter =
-            kall.vilkårperiode
-                .hentForBehandling(behandlingId)
-                .vilkårperioder.aktiviteter
-        testdataProvider.vilkår.updateDagligReise
-            .map { it(vilkårDagligReise, aktiviteter) }
-            .forEach { (vilkårId, dto) -> kall.vilkårDagligReise.oppdaterVilkår(dto, vilkårId, behandlingId) }
-
-        testdataProvider.vilkår.deleteDagligReise
-            .map { it(vilkårDagligReise) }
-            .forEach { (vilkårId, dto) -> kall.vilkårDagligReise.slettVilkår(behandlingId, vilkårId, dto) }
-    } else {
-        testdataProvider.vilkår.update
-            .map { it(vilkår) }
-            .forEach { kall.vilkår.oppdaterVilkår(it) }
-
-        testdataProvider.vilkår.delete
-            .map { it(vilkår) }
-            .forEach { kall.vilkår.slettVilkår(it) }
-    }
-
-    kall.steg.ferdigstill(
-        behandlingId,
-        StegController.FerdigstillStegRequest(
-            steg = StegType.VILKÅR,
-        ),
-    )
-    kjørTasksKlareForProsessering()
-}
